+++ dc1/net.py
@@ -1,0 +1,42 @@
+import torch
+import torch.nn as nn
+
+
+class Net(nn.Module):
+    def __init__(self, n_classes: int) -> None:
+        super(Net, self).__init__()
+
+        self.cnn_layers = nn.Sequential(
+            # Defining a 2D convolution layer
+            nn.Conv2d(1, 64, kernel_size=4, stride=1),
+            nn.BatchNorm2d(64),
+            nn.ReLU(inplace=True),
+            nn.MaxPool2d(kernel_size=4),
+            torch.nn.Dropout(p=0.5, inplace=True),
+            # Defining another 2D convolution layer
+            nn.Conv2d(64, 32, kernel_size=4, stride=1),
+            nn.BatchNorm2d(32),
+            nn.ReLU(inplace=True),
+            nn.MaxPool2d(kernel_size=3),
+            torch.nn.Dropout(p=0.25, inplace=True),
+            # Defining another 2D convolution layer
+            nn.Conv2d(32, 16, kernel_size=4, stride=1),
+            nn.BatchNorm2d(16),
+            nn.ReLU(inplace=True),
+            nn.MaxPool2d(kernel_size=2),
+            torch.nn.Dropout(p=0.125, inplace=True),
+        )
+
+        self.linear_layers = nn.Sequential(
+            nn.Linear(144, 256),
+            nn.Linear(256, n_classes)
+        )
+
+    # Defining the forward pass
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        x = self.cnn_layers(x)
+        # After our convolutional layers which are 2D, we need to flatten our
+        # input to be 1 dimensional, as the linear layers require this.
+        x = x.view(x.size(0), -1)
+        x = self.linear_layers(x)
+        return x
