+++ requirements.txt
@@ -1,7 +1,23 @@
-torch>=1.12
-torchsummary>=1.5.1
-numpy
-matplotlib
-tqdm
-requests
-plotext
+charset_normalizer==2.1.1
+certifi==2022.12.7
+colorama==0.4.6
+fonttools==4.38.0
+idna==3.4
+kiwisolver==1.4.4
+matplotlib==3.6.2
+mypy==0.991
+numpy==1.24.1
+packaging==22.0
+pillow==9.4.0
+plotext==5.2.8
+pyparsing==3.0.9
+python-dateutil==2.8.2
+requests==2.28.1
+six==1.16.0
+torch==1.13.1
+torchsummary==1.5.1
+tqdm==4.64.1
+types-requests==2.28.11.7
+types-setuptools==65.6.0.2
+types-tqdm==4.64.1
+urllib3==1.26.13
